Add histogram normalization script and dataset train/test config (#69)
* Add histogram normalization script and dataset train/test config

* update predicted dataset folder name in run.sh

* fix mention of TEST_DATA_DIR and path logic

* Comment out gpu check and make moving window forgiving of no checkpoints

* added dataset manager.

* rename from config dir to dataset dir.

* convert datasets.txt to python list.

* added image dimension checking to DatasetsManager.

* revert commenting of gpu check

* cleaned up some existing code involving enface.

* generate single enface image function.

* refactoring of dataset and config code.

* more code changes.

* path bug fix.

* simplified checkpoint cleanup code.

* minor cleanup of generate_inferred_images code.

+++ src/model_state.py
@@ -2,8 +2,9 @@ import os
 import tensorflow as tf
 from src.generator import generator
 from src.discriminator import discriminator
-from src.parameters import OUTPUT_CHANNELS, TEST_DATA_DIR, TRAIN_DATA_DIR
+from configs.parameters import OUTPUT_CHANNELS, ALL_DATA_DIR
 from src.utils import get_dataset
+from datasets.train_and_test import TRAINING_DATASETS, TESTING_DATASETS
 
 
 class ModelState:
@@ -13,44 +14,46 @@ class ModelState:
         self.generator_optimizer = tf.keras.optimizers.Adam(5e-4, beta_1=0.5)
         self.generator = generator(OUTPUT_CHANNELS)
         self.discriminator = discriminator()
-        self.test_dataset = get_dataset(TEST_DATA_DIR)
-        self.train_dataset = get_dataset(TRAIN_DATA_DIR)
+        self.test_dataset = get_dataset(ALL_DATA_DIR, dataset_list=TESTING_DATASETS)
+        self.train_dataset = get_dataset(ALL_DATA_DIR, dataset_list=TRAINING_DATASETS)
         self.checkpoint_dir = './training_checkpoints'
         self.checkpoint_prefix = os.path.join(self.checkpoint_dir, 'ckpt')
-        self.checkpoint = tf.train.Checkpoint(generator_optimizer=self.generator_optimizer,
-                                              discriminator_optimizer=self.discriminator_optimizer,
-                                              generator=self.generator,
-                                              discriminator=self.discriminator)
+        self.checkpoint = tf.train.Checkpoint(
+            generator_optimizer=self.generator_optimizer,
+            discriminator_optimizer=self.discriminator_optimizer,
+            generator=self.generator,
+            discriminator=self.discriminator
+        )
 
     def save_checkpoint(self):
         self.checkpoint.save(file_prefix=self.checkpoint_prefix)
 
     def restore_from_checkpoint(self):
-        self.__moving_window_checkpoint_cleanup()  # cleanup old checkpoints to reduce memory footprint
+        # cleanup old checkpoints to reduce memory footprint
+        self.__moving_window_checkpoint_cleanup()
         self.checkpoint.restore(tf.train.latest_checkpoint(self.checkpoint_dir))
 
+    @staticmethod
+    def __get_epoch_num(checkpoint_name):
+        return int(checkpoint_name.split(".")[0].split("-")[1])
+
     def __moving_window_checkpoint_cleanup(self, window_size=5):
-        checkpoint_files = os.listdir(self.checkpoint_dir)  # list of all checkpoint files
-        checkpoint_files.remove("checkpoint")  # no need to consider this file
+        # list of all checkpoint files
+        os.makedirs(self.checkpoint_dir, exist_ok=True)
+        checkpoint_files = os.listdir(self.checkpoint_dir)
+        if 'checkpoint' in checkpoint_files:
+            # no need to consider this file
+            checkpoint_files.remove("checkpoint")
 
         if len(checkpoint_files) == 0:
             return
 
         # sort by asc epoch numbers
-        checkpoint_files.sort(key=lambda elem: int(elem.split(".")[0].split("-")[1]), reverse=False)
+        checkpoint_files.sort(key=self.__get_epoch_num, reverse=False)
 
-        # account for multiple records e.g. ckpt-30.data-00000-of-00002, ckpt-30.data-00001-of-00002, ckpt-30.index
-        # by parsing nnnnn from `ckpt-30.data-00001-of-nnnnn` file name
-        try:
-            checkpoint_multiplier = int(checkpoint_files[-1].split(".")[1].split("-")[3]) + 1
-        except IndexError:
-            # last element is of format `ckpt-xx.index`, cannot use, fall back to second-last
-            # element of `ckpt-xx.data-xxxxx-of-xxxxx` format
-            checkpoint_multiplier = int(checkpoint_files[-2].split(".")[1].split("-")[3]) + 1
-
-        checkpoints_to_keep = checkpoint_files[-1 * window_size * checkpoint_multiplier:]  # save most recent
-        checkpoints_to_delete = list(filter(lambda elem: elem not in checkpoints_to_keep, checkpoint_files))
-
-        # delete the old checkpoints
-        for checkpoint in checkpoints_to_delete:
-            os.remove(self.checkpoint_dir + "/" + checkpoint)
+        # cleanup old checkpoints
+        highest_epoch_num = self.__get_epoch_num(checkpoint_files[-1])
+        for checkpoint in checkpoint_files:
+            if self.__get_epoch_num(checkpoint) < (highest_epoch_num - window_size):
+                os.remove(self.checkpoint_dir + "/" + checkpoint)
+                print("cleaned up checkpoint ", checkpoint)
